feat(ui): add context lib
