--- apps/dsb-client-gateway-frontend/src/context/index.ts
@@ -1,1 +1,0 @@
-export * from './Backdrop/provider';
