+++ libs/dsb-client-gateway/ui/context/jest.config.js
@@ -1,0 +1,9 @@
+module.exports = {
+  displayName: 'dsb-client-gateway-ui-context',
+  preset: '../../../../jest.preset.js',
+  transform: {
+    '^.+\\.[tj]sx?$': 'babel-jest',
+  },
+  moduleFileExtensions: ['ts', 'tsx', 'js', 'jsx'],
+  coverageDirectory: '../../../../coverage/libs/dsb-client-gateway/ui/context',
+};
