+++ libs/dsb-client-gateway/ui/login/src/lib/LoginStatus/LoginStatus.tsx
@@ -1,21 +1,14 @@
 import LoadingInfo from '../LoadingInfo/LoadingInfo';
 import { useLoginStatusEffects } from './RequestingEnrolment/LoginStatus.effects';
-import { Backdrop } from "@dsb-client-gateway/ui/core";
-import { useBackdropContext } from "../../../../../../../apps/dsb-client-gateway-frontend/src/context"
 
 export function LoginStatus() {
   const { isLoading, statusFactory } = useLoginStatusEffects();
-  const { open, setLoader } = useBackdropContext();
 
-  return (
-    <>
-      {isLoading ? (
-        <Backdrop open={open} setLoader={setLoader}><LoadingInfo>Checking identity</LoadingInfo></Backdrop>
-      ) : (
-        statusFactory()
-      )}
-    </>
-  );
+  return isLoading ? (
+    <LoadingInfo>Checking identity</LoadingInfo>
+  ) : (
+    statusFactory()
+  )
 }
 
 export default LoginStatus;
