+++ tsconfig.base.json
@@ -62,6 +62,9 @@
       "@dsb-client-gateway/dsb-client-gateway/identity/models": [
         "libs/dsb-client-gateway/identity/models/src/index.ts"
       ],
+      "@dsb-client-gateway/ui/context": [
+        "libs/dsb-client-gateway/ui/context/src/index.ts"
+      ],
       "@dsb-client-gateway/ui/api-hooks": [
         "libs/dsb-client-gateway/ui/api-hooks/src/index.ts"
       ],
