+++ workspace.json
@@ -23,6 +23,7 @@
     "dsb-client-gateway-ui-api-hooks": "libs/dsb-client-gateway/ui/api-hooks",
     "dsb-client-gateway-ui-applications": "libs/dsb-client-gateway/ui/applications",
     "dsb-client-gateway-ui-channels": "libs/dsb-client-gateway/ui/channels",
+    "dsb-client-gateway-ui-context": "libs/dsb-client-gateway/ui/context",
     "dsb-client-gateway-ui-integration": "libs/dsb-client-gateway/ui/integration",
     "dsb-client-gateway-ui-login": "libs/dsb-client-gateway/ui/login",
     "dsb-client-gateway-ui-messaging": "libs/dsb-client-gateway/ui/messaging",
